Removed Redudant Class: Base Menu Made the Title Screen to a functional degree.

+++ ERPOpgave/ERPOpgave/Order/Orderline.cs
@@ -8,59 +8,17 @@ namespace ERPOpgave.Order
 {
     internal class Orderline
     {
-        public int ItemNumber
-        {
-            get => default;
-            set
-            {
-            }
-        }
+        public int ItemNumber { get => default; set { } }
 
-        public string ItemText
-        {
-            get => default;
-            set
-            {
-            }
-        }
+        public string ItemText { get => default; set { } }
 
-        public int Property
-        {
-            get => default;
-            set
-            {
-            }
-        }
+        public int Property { get => default; set { } }
 
-        public int Quantity
-        {
-            get => default;
-            set
-            {
-            }
-        }
+        public int Quantity { get => default; set { } }
 
-        public int Units
-        {
-            get => default;
-            set
-            {
-            }
-        }
+        public int Units { get => default; set { } }
 
-        public int UnitPrice
-        {
-            get => default;
-            set
-            {
-            }
-        }
-        internal Products.Product Product
-        {
-            get => default;
-            set
-            {
-            }
-        }
+        public decimal UnitPrice { get => default; set { } }
+        internal Products.Product Product { get => default; set { } }
     }
 }
